requirements.txt
pynostr==0.6.2
setuptools==67.3.2
cachetools==5.3.0
cryptography==39.0.1
secp256k1==0.14.0
bech32==1.2.0
python-dotenv